shawn/requirements.txt
appdirs==1.4.4
azure-common==1.1.26
azure-core==1.10.0
azure-storage-blob==2.1.0
azure-storage-common==2.1.0
Brotli==1.0.9
certifi==2020.12.5
cffi==1.14.4
chardet==4.0.0
click==7.1.2
cryptography==3.3.1
dash==1.18.1
dash-bootstrap-components==0.11.3
dash-core-components==1.14.1
dash-html-components==1.1.1
dash-renderer==1.8.3
dash-table==4.11.1
distlib==0.3.1
filelock==3.0.12
Flask==1.1.2
Flask-Compress==1.8.0
future==0.18.2
idna==2.10
isodate==0.6.0
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
msrest==0.6.19
numpy==1.20.1
oauthlib==3.1.0
pandas==1.2.0
plotly==4.14.1
pycparser==2.20
python-dateutil==2.8.1
pytz==2020.5
requests==2.25.1
requests-oauthlib==1.3.0
retrying==1.3.3
six==1.15.0
urllib3==1.26.2
virtualenv==20.2.2
Werkzeug==1.0.1
